+++ vaultaire-example/src/test/kotlin/com/github/manosbatsis/vaultaire/example/FlowTests.kt
@@ -34,6 +34,7 @@ import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
+import java.math.BigDecimal
 import kotlin.test.assertEquals
 
 
@@ -65,19 +66,33 @@ class FlowTests {
         network.stopNodes()
     }
 
-    @Test
-    fun `Test generated conditions DSL `() {
-        val bookTitle = "A bok on Corda"
+    //@Test
+    fun `Test DSL conditions`() {
+        val bookTitle = "A book on Corda"
         val bookState = BookContract.BookState(
                 publisher = a.info.legalIdentities.first(),
                 author = b.info.legalIdentities.first(),
+                price = BigDecimal.valueOf(10),
+                genre = BookContract.BookGenre.TECHNOLOGY,
+                editions = 3,
                 title = bookTitle)
+
+
 
         val stx = flowWorksCorrectly(
                 CreateBookFlow(BookMessage(
                         author = b.info.legalIdentities.first(),
+                        price = BigDecimal.valueOf(10),
+                        genre = BookContract.BookGenre.TECHNOLOGY,
+                        editions = 3,
                         title = bookTitle)))
-
+        flowWorksCorrectly(
+                CreateBookFlow(BookMessage(
+                        author = b.info.legalIdentities.first(),
+                        price = BigDecimal.valueOf(20),
+                        genre = BookContract.BookGenre.TECHNOLOGY,
+                        editions = 1,
+                        title = "$bookTitle 2")))
         // Check book transaction is stored in the storage service.
         val bTx = b.services.validatedTransactions.getTransaction(stx.id)
         assertEquals(bTx, stx)
@@ -101,44 +116,123 @@ class FlowTests {
                 and {
                     fields.title `==` bookTitle
                     or {
-                        fields.title `!=` "false match"
-                        //fields.title notEqual bookTitle
+                        fields.title notEqual bookTitle
                         fields["title"] _notEqual bookTitle
+                        fields.title `!=` "false match"
+                    }
+                    or {
+                        fields.description.isNull()
+                        fields.description.notNull()
                     }
                 }
                 orderBy {
                     fields.title sort ASC
+                    fields.description sort ASC
                 }
             }
 
-
             // Test BasicStateService
             val stateBasicService = BasicStateService(b.services, BookContract.BookState::class.java)
-            testStateService(stateBasicService, bookStateQuery, bookState)
+            testStateServiceQueryBy(stateBasicService, bookStateQuery, bookState)
 
             // Test manually coded subclass of BasicStateService
             val bookBasicStateService = CustomBasicBookStateService(b.services)
-            testStateService(bookBasicStateService, bookStateQuery, bookState)
+            testStateServiceQueryBy(bookBasicStateService, bookStateQuery, bookState)
 
             // Test generated BookStateService
             val serviceHubBookStateService = BookStateService(b.services)
-            testStateService(bookBasicStateService, bookStateQuery, bookState)
+            testStateServiceQueryBy(bookBasicStateService, bookStateQuery, bookState)
 
             // Test manually coded subclass of BookStateService
             val myBookStateService = MyExtendedBookStateService(b.services)
-            testStateService(myBookStateService, bookStateQuery, bookState)
+            testStateServiceQueryBy(myBookStateService, bookStateQuery, bookState)
 
         }
     }
 
-    private fun testStateService(stateService: BasicStateService<BookContract.BookState>, bookStateQuery: PersistentBookStateConditions, bookState: BookContract.BookState) {
-        var bookSearchResult = stateService.queryBy(
-                bookStateQuery.toCriteria(), 1, 1, bookStateQuery.toSort()
-        ).states.single().state.data
+    @Test
+    fun `Test DSL aggregations `() {
+
+        val stx = flowWorksCorrectly(
+                CreateBookFlow(BookMessage(
+                        author = b.info.legalIdentities.first(),
+                        price = BigDecimal.valueOf(8),
+                        genre = BookContract.BookGenre.FANTACY,
+                        editions = 3,
+                        title = "Lord of Corda")))
+        flowWorksCorrectly(
+                CreateBookFlow(BookMessage(
+                        author = b.info.legalIdentities.first(),
+                        price = BigDecimal.valueOf(10),
+                        genre = BookContract.BookGenre.FANTACY,
+                        editions = 2,
+                        title = "Lord of Corda 2")))
+        flowWorksCorrectly(
+                CreateBookFlow(BookMessage(
+                        author = b.info.legalIdentities.first(),
+                        price = BigDecimal.valueOf(12),
+                        genre = BookContract.BookGenre.FANTACY,
+                        editions = 2,
+                        title = "Lord of Corda 3")))
+
+
+        // Use the generated DSL to create query criteria
+        val bookStateQuery = bookStateQuery {
+            status = Vault.StateStatus.UNCONSUMED // the default
+            relevancyStatus = Vault.RelevancyStatus.ALL // the default
+            and {
+                fields.title `like` "Lord of Corda%"
+                fields.genre `==` BookContract.BookGenre.FANTACY
+
+                // add some aggregates
+                fields.externalId.count()
+                fields.id.count()
+                fields.editions.sum()
+                fields.price.min()
+                fields.price.avg()
+                fields.price.max()
+            }
+            orderBy {
+                fields.title sort ASC
+            }
+        }
+
+
+        // Test BasicStateService
+        val stateBasicService = BasicStateService(b.services, BookContract.BookState::class.java)
+
+        val stateService = CustomBasicBookStateService(b.services)
+        val aggregateResults = stateService.queryBy(bookStateQuery.toCriteria())
+        aggregateResults.otherResults.forEachIndexed { index, element ->
+            println("testStateServiceAggregates, aggregate $index: $element")
+        }
+        // Must be five aggregates
+        assertEquals(6, aggregateResults.otherResults.size)
+        // Must zero external IDs
+        assertEquals(0.toLong(), aggregateResults.otherResults[0])
+        // Must three linear IDs
+        assertEquals(3.toLong(), aggregateResults.otherResults[1])
+        // Must be seven editions
+        assertEquals(7.toLong(), aggregateResults.otherResults[2])
+        // Minimum price must be eight
+        assertEquals(0, BigDecimal(8).compareTo(aggregateResults.otherResults[3] as BigDecimal))
+        // Average price must be ten
+        assertEquals(0, BigDecimal.TEN.compareTo(BigDecimal(aggregateResults.otherResults[4] as Double)))
+        // Minimum price must be 12
+        assertEquals(0, BigDecimal(12).compareTo(aggregateResults.otherResults[5] as BigDecimal))
+    }
+
+    private fun testStateServiceQueryBy(stateService: BasicStateService<BookContract.BookState>, bookStateQuery: PersistentBookStateConditions, bookState: BookContract.BookState) {
+        val bookSearchPage = stateService.queryBy(
+                bookStateQuery.toCriteria(true), 1, 10, bookStateQuery.toSort()
+        )
+        val bookSearchResult = bookSearchPage.states.single().state.data
         assertEquals(bookState.title, bookSearchResult.title)
         assertEquals(1, stateService.countBy(bookStateQuery.toCriteria()))
         print("$bookSearchResult == $bookState\n")
     }
+
+
 
 
     inline fun <reified OUT> flowWorksCorrectly(flow: PartitureFlow<*, OUT>): OUT {