+++ vaultaire/src/main/kotlin/com/github/manosbatsis/vaultaire/dsl/ConditionsDsl.kt
@@ -21,12 +21,16 @@ package com.github.manosbatsis.vaultaire.dsl
 
 import com.github.manosbatsis.vaultaire.util.FieldWrapper
 import com.github.manosbatsis.vaultaire.util.Fields
+import com.github.manosbatsis.vaultaire.util.NullableGenericFieldWrapper
+import com.github.manosbatsis.vaultaire.util.TypedFieldWrapper
 import net.corda.core.contracts.ContractState
 import net.corda.core.contracts.StateRef
 import net.corda.core.identity.AbstractParty
 import net.corda.core.node.services.Vault
 import net.corda.core.node.services.vault.Builder.`in`
+import net.corda.core.node.services.vault.Builder.avg
 import net.corda.core.node.services.vault.Builder.between
+import net.corda.core.node.services.vault.Builder.count
 import net.corda.core.node.services.vault.Builder.equal
 import net.corda.core.node.services.vault.Builder.greaterThan
 import net.corda.core.node.services.vault.Builder.greaterThanOrEqual
@@ -34,10 +38,13 @@ import net.corda.core.node.services.vault.Builder.isNull
 import net.corda.core.node.services.vault.Builder.lessThan
 import net.corda.core.node.services.vault.Builder.lessThanOrEqual
 import net.corda.core.node.services.vault.Builder.like
+import net.corda.core.node.services.vault.Builder.max
+import net.corda.core.node.services.vault.Builder.min
 import net.corda.core.node.services.vault.Builder.notEqual
 import net.corda.core.node.services.vault.Builder.notIn
 import net.corda.core.node.services.vault.Builder.notLike
 import net.corda.core.node.services.vault.Builder.notNull
+import net.corda.core.node.services.vault.Builder.sum
 import net.corda.core.node.services.vault.QueryCriteria
 import net.corda.core.node.services.vault.QueryCriteria.VaultCustomQueryCriteria
 import net.corda.core.node.services.vault.Sort
@@ -51,7 +58,7 @@ import kotlin.Suppress as supress
 interface Condition {
 
     /** Implement to expose internal state as criteria */
-    abstract fun toCriteria(): QueryCriteria?
+    fun toCriteria(ignoreAggregates: Boolean = false): QueryCriteria?
 }
 
 /** A [Condition] that contains other conditions. Allows for nested and/or condition groups */
@@ -61,9 +68,14 @@ abstract class ConditionsCondition<P : StatePersistable, out F: Fields<P>>: Cond
     abstract val fields: F
 
     val conditions: MutableList<Condition> = mutableListOf()
+    val aggregates: MutableList<Condition> = mutableListOf()
 
     fun addCondition(condition: Condition) {
         conditions += condition
+    }
+
+    fun addAggregate(condition: Condition) {
+        aggregates += condition
     }
 
     fun and(initializer: CompositeCondition<P, F>.() -> Unit) = addCondition(AndCondition(fields).apply(initializer))
@@ -80,87 +92,112 @@ abstract class CompositeCondition<P : StatePersistable, out F: Fields<P>>(
         override val fields: F
 ) : ConditionsCondition<P, F>() {
 
-    fun <S : Comparable<S>> FieldWrapper<P, S>.isNull() =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.isNull())))
+    fun <S> NullableGenericFieldWrapper<P, S>.isNull() =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.isNull())))
 
-    fun <S : Comparable<S>> FieldWrapper<P, S>.notNull() =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.notNull())))
+    fun <S> NullableGenericFieldWrapper<P, S>.notNull() =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.notNull())))
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.equal(value: S) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.equal(value))))
+    infix fun <S> TypedFieldWrapper<P, S>.equal(value: S) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.equal(value))))
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.`==`(value: S) = equal(value)
+    infix fun <S> TypedFieldWrapper<P, S>.`==`(value: S) = equal(value)
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.notEqual(value: S) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.notEqual(value))))
+    infix fun <S> TypedFieldWrapper<P, S>.notEqual(value: S) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.notEqual(value))))
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.`!=`(value: S) = notEqual(value)
+    infix fun <S> TypedFieldWrapper<P, S>.`!=`(value: S) = notEqual(value)
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.lessThan(value: S) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.lessThan(value))))
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.lessThan(value: S) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.lessThan(value))))
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.lt(value: S) = lessThan(value)
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.lt(value: S) = lessThan(value)
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.lessThanOrEqual(value: S) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.lessThanOrEqual(value))))
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.lessThanOrEqual(value: S) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.lessThanOrEqual(value))))
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.lte(value: S) = lessThanOrEqual(value)
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.lte(value: S) = lessThanOrEqual(value)
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.greaterThan(value: S) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.greaterThan(value))))
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.greaterThan(value: S) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.greaterThan(value))))
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.gt(value: S) = greaterThan(value)
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.gt(value: S) = greaterThan(value)
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.greaterThanOrEqual(value: S) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.greaterThanOrEqual(value))))
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.greaterThanOrEqual(value: S) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.greaterThanOrEqual(value))))
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.gte(value: S) = greaterThanOrEqual(value)
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.gte(value: S) = greaterThanOrEqual(value)
 
-    fun <S : Comparable<S>> FieldWrapper<P, S>.between(from: S, to: S) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.between(from, to))))
+    fun <S : Comparable<S>> TypedFieldWrapper<P, S>.between(from: S, to: S) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.between(from, to))))
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.between(value: Pair<S, S>) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.between(value.first, value.second))))
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.between(value: Pair<S, S>) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.between(value.first, value.second))))
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.btw(value: Pair<S, S>) = between(value)
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.btw(value: Pair<S, S>) = between(value)
 
-    infix fun <T : StatePersistable> FieldWrapper<T, String>.like(value: String) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.like(value))))
+    infix fun <T : StatePersistable> TypedFieldWrapper<T, String>.like(value: String) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.like(value))))
 
-    fun <T : StatePersistable> FieldWrapper<T, String>.like(value: String, exactMatch: Boolean = true) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.like(value, exactMatch))))
+    fun <T : StatePersistable> TypedFieldWrapper<T, String>.like(value: String, exactMatch: Boolean = true) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.like(value, exactMatch))))
 
-    infix fun <T : StatePersistable> FieldWrapper<T, String>.notLike(value: String) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.notLike(value))))
+    infix fun <T : StatePersistable> TypedFieldWrapper<T, String>.notLike(value: String) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.notLike(value))))
 
-    fun <T : StatePersistable> FieldWrapper<T, String>.notLike(value: String, exactMatch: Boolean = true) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.notLike(value, exactMatch))))
+    fun <T : StatePersistable> TypedFieldWrapper<T, String>.notLike(value: String, exactMatch: Boolean = true) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.notLike(value, exactMatch))))
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.isIn(value: Collection<S>) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.`in`(value))))
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.isIn(value: Collection<S>) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.`in`(value))))
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.`in`(value: Collection<S>) = isIn(value)
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.`in`(value: Collection<S>) = isIn(value)
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.notIn(value: Collection<S>) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.notIn(value))))
+    infix fun <S : Comparable<S>> TypedFieldWrapper<P, S>.notIn(value: Collection<S>) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.notIn(value))))
 
-    fun <S : Comparable<S>> FieldWrapper<P, S>.`!in`(value: Collection<S>) = notIn(value)
+    fun <S : Comparable<S>> TypedFieldWrapper<P, S>.`!in`(value: Collection<S>) = notIn(value)
 
     // non type safe versions
-    infix fun FieldWrapper<P, *>._equal(value: Any) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.equal(value))))
+    infix fun FieldWrapper<P>._equal(value: Any) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.equal(value))))
 
-    infix fun FieldWrapper<P, *>._notEqual(value: Any) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.notEqual(value))))
+    infix fun FieldWrapper<P>._notEqual(value: Any) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.property.notEqual(value))))
 
-    infix fun FieldWrapper<P, *>._like(value: String) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.asStringProperty().like(value))))
+    infix fun FieldWrapper<P>._like(value: String) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.asStringProperty().like(value))))
 
-    infix fun FieldWrapper<P, *>._notLike(value: String) =
-            addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.asStringProperty().notLike(value))))
+    infix fun FieldWrapper<P>._notLike(value: String) =
+            addCondition(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(this.asStringProperty().notLike(value))))
 
     @supress("UNCHECKED_CAST")
-    protected fun FieldWrapper<P, *>.asStringProperty() = this.property as KProperty1<P, String>
+    protected fun FieldWrapper<P>.asStringProperty() = this.property as KProperty1<P, String>
+
+    // aggregates
+    fun <S : Comparable<S>> TypedFieldWrapper<P, S>.sum(
+            groupColumns: List<FieldWrapper<P>>? = null, orderBy: Sort.Direction? = null) =
+            addAggregate(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(
+                    this.property.sum(if(groupColumns != null) groupColumns.map { it.property } else null, orderBy))))
+
+    fun <S : Comparable<S>> TypedFieldWrapper<P, S>.avg(
+            groupColumns: List<FieldWrapper<P>>? = null, orderBy: Sort.Direction? = null) =
+            addAggregate(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(
+                    this.property.avg(if(groupColumns != null) groupColumns.map { it.property } else null, orderBy))))
+
+    fun <S : Comparable<S>> TypedFieldWrapper<P, S>.min(
+            groupColumns: List<FieldWrapper<P>>? = null, orderBy: Sort.Direction? = null) =
+            addAggregate(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(
+                    this.property.min(if(groupColumns != null) groupColumns.map { it.property } else null, orderBy))))
+
+    fun <S : Comparable<S>> TypedFieldWrapper<P, S>.max(
+            groupColumns: List<FieldWrapper<P>>? = null, orderBy: Sort.Direction? = null) =
+            addAggregate(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(
+                    this.property.max(if(groupColumns != null) groupColumns.map { it.property } else null, orderBy))))
+
+    fun FieldWrapper<P>.count() =
+            addAggregate(VaultCustomQueryCriteriaWrapperCondition(VaultCustomQueryCriteria(
+                    this.property.count())))
 }
 
 /** Defines a set of conditions where all items must be matched */
@@ -168,10 +205,12 @@ class AndCondition<P : StatePersistable, out F: Fields<P>>(
         fields: F
 ) : CompositeCondition<P, F>(fields) {
 
-    override fun toCriteria(): QueryCriteria? {
+
+    override fun toCriteria(ignoreAggregates: Boolean): QueryCriteria? {
         var criteria: QueryCriteria? = null
-        this.conditions.forEach {
-            val childCriteria = it.toCriteria()
+        val conditionsToProcess = if(ignoreAggregates) this.conditions else this.conditions.plus(this.aggregates)
+        conditionsToProcess.forEach {
+            val childCriteria = it.toCriteria(ignoreAggregates)
             if(childCriteria != null) {
                 if(criteria == null ) criteria = childCriteria
                 else criteria = criteria!!.and(childCriteria)
@@ -186,13 +225,20 @@ class OrCondition<P : StatePersistable, out F: Fields<P>>(
         fields: F
 ) : CompositeCondition<P, F>(fields) {
 
-    override fun toCriteria(): QueryCriteria? {
+    override fun toCriteria(ignoreAggregates: Boolean): QueryCriteria? {
         var criteria: QueryCriteria? = null
         this.conditions.forEach {
-            val childCriteria = it.toCriteria()
+            val childCriteria = it.toCriteria(ignoreAggregates)
             if(childCriteria != null) {
                 if(criteria == null ) criteria = childCriteria
                 else criteria = criteria!!.or(childCriteria)
+            }
+        }
+        if(!ignoreAggregates) aggregates.forEach {
+            val childCriteria = it.toCriteria(ignoreAggregates)
+            if(childCriteria != null) {
+                if(criteria == null ) criteria = childCriteria
+                else criteria = criteria!!.and(childCriteria)
             }
         }
         return criteria
@@ -200,12 +246,16 @@ class OrCondition<P : StatePersistable, out F: Fields<P>>(
 }
 
 /**
- * A condition equivalent to a [VaultCustomQueryCriteria] instance.
- * Created by infix and regular extension functions of [FieldWrapper]
- * defined by [CompositeCondition]
+ * A condition that wraps a [VaultCustomQueryCriteria] instance.
+ * Instances of this type are typically created by
+ * [FieldWrapper] extension functions defined by [CompositeCondition]
+ * as shorthands (infix or regular), i.e. expression-only criteria without a contract state type,
+ * vault state status or relevance status of their own.
  */
-class VaultCustomQueryCriteriaCondition(val criterion: VaultCustomQueryCriteria<*>) : Condition {
-    override fun toCriteria(): QueryCriteria = criterion
+open class VaultCustomQueryCriteriaWrapperCondition(
+        private val criterion: VaultCustomQueryCriteria<*>
+) : Condition {
+    final override fun toCriteria(ignoreAggregates: Boolean): QueryCriteria = criterion
 }
 
 /** Used to define [Sort.SortColumn]s */
@@ -215,10 +265,11 @@ class SortColumns<P : StatePersistable>(val statePersistableType: Class<P>){
     val DESC = Sort.Direction.DESC
     val entries: LinkedHashSet<Sort.SortColumn> = linkedSetOf()
 
-    infix fun <S : Comparable<S>> FieldWrapper<P, S>.sort(value: Sort.Direction) {
+    infix fun FieldWrapper<P>.sort(value: Sort.Direction) {
         entries.add(Sort.SortColumn(SortAttribute.Custom(statePersistableType, this.property.name), value))
     }
 }
+
 
 /**
  * A [ConditionsCondition] extended by Vaultaire's annotation processing to create a condition DSL specific to a [ContractState] type.
@@ -250,13 +301,13 @@ abstract class VaultQueryCriteriaCondition<P : StatePersistable, out F: Fields<P
 
     fun toSort(): Sort = Sort(sortColumns.entries)
 
-    override fun toCriteria(): QueryCriteria {
+    override fun toCriteria(ignoreAggregates: Boolean): QueryCriteria {
         var criteria: QueryCriteria = QueryCriteria.VaultQueryCriteria(
                 status, setOf(contractStateType), stateRefs, notary, softLockingCondition, timeCondition,
                 relevancyStatus, constraintTypes, constraints, participants)
 
         this.conditions.forEach {
-            val childCriteria = it.toCriteria()
+            val childCriteria = it.toCriteria(ignoreAggregates)
             if(childCriteria != null) criteria = criteria.and(childCriteria)
         }